# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Boilerplate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  code?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
  totpCode?: string;
}

export interface RefreshTokenBody {
  refreshToken: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  /** @minLength 8 */
  password: string;
}

export interface VerifyEmailBody {
  token: string;
}

export interface TwoFactorSetupResponse {
  secret: string;
  qrCodeUrl: string;
  backupCodes: string[];
}

export interface TwoFactorVerifyBody {
  totpCode: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  SUPER_ADMIN: "SUPER_ADMIN",
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export type UserProfilePlan =
  (typeof UserProfilePlan)[keyof typeof UserProfilePlan];

export const UserProfilePlan = {
  FREE: "FREE",
  PRO: "PRO",
  ENTERPRISE: "ENTERPRISE",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  name: string;
  role: UserProfileRole;
  plan: UserProfilePlan;
  emailVerified: boolean;
  twoFactorEnabled: boolean;
  avatarUrl?: string | null;
  createdAt: string;
  banned: boolean;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: UserProfile;
}

export interface UpdateProfileBody {
  name?: string;
  avatarUrl?: string;
}

export interface ChangePasswordBody {
  oldPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface Session {
  id: string;
  ipAddress?: string | null;
  userAgent?: string | null;
  createdAt: string;
  expiresAt: string;
}

export type CheckoutBodyPlan =
  (typeof CheckoutBodyPlan)[keyof typeof CheckoutBodyPlan];

export const CheckoutBodyPlan = {
  PRO: "PRO",
  ENTERPRISE: "ENTERPRISE",
} as const;

export interface CheckoutBody {
  plan: CheckoutBodyPlan;
}

export interface CheckoutResponse {
  url: string;
}

export interface PortalResponse {
  url: string;
}

export type SubscriptionPlan =
  (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan];

export const SubscriptionPlan = {
  FREE: "FREE",
  PRO: "PRO",
  ENTERPRISE: "ENTERPRISE",
} as const;

export interface Subscription {
  id?: string;
  plan: SubscriptionPlan;
  status: string;
  currentPeriodStart?: string | null;
  currentPeriodEnd?: string | null;
  cancelAt?: string | null;
}

export interface Invoice {
  id: string;
  amount: number;
  currency: string;
  status: string;
  createdAt: string;
  pdfUrl?: string | null;
}

export interface InvoiceListResponse {
  data: Invoice[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardKpis {
  totalUsers: number;
  totalUsersChange: number;
  mrr: number;
  mrrChange: number;
  dau: number;
  dauChange: number;
  apiCallsThisMonth: number;
  apiCallsLimit: number;
  churnRate: number;
  arpu: number;
}

export interface TimeSeriesPoint {
  date: string;
  value: number;
}

export interface PlanDistribution {
  free: number;
  pro: number;
  enterprise: number;
}

export interface ActivityEvent {
  id: string;
  action: string;
  userId: string;
  userName: string;
  userEmail: string;
  resourceType?: string | null;
  ipAddress?: string | null;
  createdAt: string;
}

export interface EndpointStat {
  endpoint: string;
  calls: number;
  avgLatency: number;
}

export interface ApiUsageStats {
  totalCalls: number;
  successRate: number;
  p50Latency: number;
  p95Latency: number;
  p99Latency: number;
  errorRate: number;
  topEndpoints: EndpointStat[];
}

export interface AdminUserListResponse {
  data: UserProfile[];
  total: number;
  page: number;
  limit: number;
}

export type ChangeRoleBodyRole =
  (typeof ChangeRoleBodyRole)[keyof typeof ChangeRoleBodyRole];

export const ChangeRoleBodyRole = {
  SUPER_ADMIN: "SUPER_ADMIN",
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export interface ChangeRoleBody {
  role: ChangeRoleBodyRole;
}

export type GetInvoicesParams = {
  page?: number;
  limit?: number;
};

export type GetAdminUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
  plan?: string;
  role?: string;
};